+++ packages/block-library/src/block/edit.js
@@ -29,6 +29,7 @@ import {
 	privateApis as blockEditorPrivateApis,
 	store as blockEditorStore,
 	BlockControls,
+	InnerBlocks,
 } from '@wordpress/block-editor';
 import { privateApis as patternsPrivateApis } from '@wordpress/patterns';
 import { getBlockBindingsSource } from '@wordpress/blocks';
@@ -218,6 +219,9 @@ function ReusableBlockEdit( {
 		value: blocks,
 		onInput: NOOP,
 		onChange: NOOP,
+		renderAppender: blocks?.length
+			? undefined
+			: InnerBlocks.ButtonBlockAppender,
 	} );
 
 	const handleEditOriginal = () => {