+++ packages/block-library/src/block/index.js
@@ -20,7 +20,6 @@ export { metadata, name };
 
 export const settings = {
 	deprecated,
-	appender: 'button',
 	edit,
 	icon,
 	__experimentalLabel: ( { ref } ) => {