+++ packages/block-library/src/column/index.js
@@ -18,7 +18,6 @@ export { metadata, name };
 
 export const settings = {
 	icon,
-	appender: 'button',
 	edit,
 	save,
 	deprecated,