+++ packages/block-library/src/form-submission-notification/index.js
@@ -31,7 +31,6 @@ const TEMPLATE = [
 export const settings = {
 	icon,
 	template: TEMPLATE,
-	appender: 'button',
 	edit,
 	save,
 	variations,