+++ packages/block-library/src/form/index.js
@@ -56,7 +56,6 @@ const TEMPLATE = [
 export const settings = {
 	icon,
 	template: TEMPLATE,
-	appender: 'button',
 	edit,
 	save,
 	deprecated,