+++ packages/block-library/src/group/index.js
@@ -79,7 +79,6 @@ export const settings = {
 		viewportWidth: 600,
 	},
 	transforms,
-	appender: 'button',
 	edit,
 	save,
 	deprecated,