+++ packages/block-library/src/quote/index.js
@@ -38,6 +38,7 @@ export const settings = {
 	transforms,
 	template: TEMPLATE,
 	templateInsertUpdatesSelection: true,
+	appender: 'default',
 	edit,
 	save,
 	deprecated,